Now there is only 1 generic RADIUS Client and Server, which could be used to build Sync/Async or whatever custom implementations one would want
