+++ src/client/mod.rs
@@ -1,6 +1,10 @@
 //! RADIUS Client implementation
 //!
-//! client - generic version of RADIUS Client, that could be used to build Async or Sync Client
+//! Client           - generic version of RADIUS Client, that could be used to build Async or Sync Client
+//! SyncClientTrait  - if you are planning to build Sync RADIUS Client, then you would need to
+//! implement this trait 
+//! AsyncClientTrait - if you are planning to build Async RADIUS Client, then you would need to
+//! implement this trait
 
 use crate::protocol::radius_packet::RadiusPacket;
 use crate::protocol::error::RadiusError;
