+++ tests/test_async_client.rs
@@ -1,7 +1,7 @@
 #![cfg(all(feature = "async-radius"))]
 
 
-use radius_rust::clients::{ client::Client, AsyncClientTrait };
+use radius_rust::client::{ client::Client, AsyncClientTrait };
 use radius_rust::protocol::dictionary::Dictionary;
 use radius_rust::protocol::error::RadiusError;
 use radius_rust::protocol::radius_packet::{ RadiusPacket, RadiusMsgType };
